Changed BukkitWorld in hopes that it will fix the error regarding ItemStack.getDamage().

+++ src/com/sk89q/worldedit/bukkit/BukkitWorld.java
@@ -415,7 +415,7 @@ public class BukkitWorld extends LocalWorld {
                 contents[i] = new BaseItemStack(
                         bukkitStack.getTypeId(),
                         bukkitStack.getAmount(), 
-                        bukkitStack.getDamage());
+                        (byte)bukkitStack.getDamage());
             }
         }
